bin/bmark_pq: make `--help` and `-?` options work
Also, stop advertising the `-h` synonym.  I find it bletcherous to steal one
of the precious single-letter options for something that has at least two
better spellings anyway.  Failing to advertise its existence means that
reusing it could at least be considered if a better use is found for this
option in the future.