Delete run_all_benchmarks()
Incredibly, despite the fact that the tests and documentation get more
complicated, and the fact that the loop is repeated in several callers, this
actually gives us less code overall.  So I think it's a win on that basis
alone.  Furthermore, as we mine bin/bmark_pq and bin/cmp_pq for extractable
similarities, I think it'll be awkward to embed that top-level loop in an
exported function, rather than letting callers loop over the right things
directly.