Delete dead code
The run_workload() function in Benchmark::PriorityQueue no longer has any
non-test callers, so delete it.

With that done, Benchmark::PriorityQueue::Shim#print_benchmark is unused;
deleting that leaves #timed_out unused in turn; and with that gone, it no
longer needs a `timeout` attribute.

Finally, removal of all of those means that we no longer depend on DateTime
or DateTime::Duration.